requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pytest
httpx
uvicorn[standard]
twilio
APScheduler==3.11.0
pytz
